requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backports.zoneinfo==0.2.1
binance-connector==1.18.0
binance-futures-connector==3.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.3
dateparser==1.1.3
frozenlist==1.3.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
multidict==6.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfiglet==0.8.post1
pyOpenSSL==22.1.0
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.28.1
service-identity==21.1.0
six==1.16.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websockets==10.4
yarl==1.8.1
zope.interface==5.5.0
